[FIX] int for loop to prevent float

+++ includes/config.py
@@ -4,7 +4,7 @@ import socket
 
 hostname = socket.gethostname()
 
-from includes.config_utils import *
+from config_utils import *
 
 create_data_path((""))
 file_handler = logging.FileHandler(filename=os.path.join("logs", "data.log"))
@@ -33,6 +33,6 @@ if FULLY_SYNCED_NOTIFICATIONS_PER_DAY > 0:
     FULLY_SYNCED_NOTIFICATIONS = True
     # (mins in a day / RUN_EVERY_X_MINUTES) = number of loops in 24 hours / FULLY_SYNCED_NOTIFICATIONS_PER_DAY
     # = STATUS to be performed on a multiple of this number
-    STATUS_NOTIFICATION_LOOP_COUNT = (
-        1440 / RUN_EVERY_X_MINUTES
-    ) / FULLY_SYNCED_NOTIFICATIONS_PER_DAY  # 1440 minutes per day #
+    STATUS_NOTIFICATION_LOOP_COUNT = int(
+        (1440 / RUN_EVERY_X_MINUTES) / FULLY_SYNCED_NOTIFICATIONS_PER_DAY
+    )  # 1440 minutes per day #
